+++ drift/test/mocks/mocks.dart
@@ -1,0 +1,10 @@
+import 'package:drift/drift.dart';
+import 'package:mockito/annotations.dart';
+import 'package:sentry/sentry.dart';
+
+@GenerateMocks([
+  Hub,
+  LazyDatabase,
+  TransactionExecutor,
+])
+void main() {}